+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## v1.0.1
 
-- Same as v1.0.0 but re-published due to a bug.
+-   Same as v1.0.0 but re-published due to a bug.
 
 ## v1.0.0
 
-- First release.
+-   First release.